springboot 1.5.8 -> 2.1.2

+++ src/main/java/uzblog/modules/blog/dao/impl/PostDaoImpl.java
@@ -55,7 +55,7 @@ public class PostDaoImpl implements PostDaoCustom {
 				.matching(q).createQuery();
 
 		org.hibernate.search.jpa.FullTextQuery query = fullTextSession.createFullTextQuery(luceneQuery);
-		query.setFirstResult(pageable.getOffset());
+		query.setFirstResult((int)pageable.getOffset());
 		query.setMaxResults(pageable.getPageSize());
 
 	    StandardAnalyzer standardAnalyzer = new StandardAnalyzer();
@@ -97,7 +97,7 @@ public class PostDaoImpl implements PostDaoCustom {
 		luceneQuery = term.createQuery();
 
 		org.hibernate.search.jpa.FullTextQuery query = fullTextSession.createFullTextQuery(luceneQuery);
-	    query.setFirstResult(pageable.getOffset());
+	    query.setFirstResult((int)pageable.getOffset());
 	    query.setMaxResults(pageable.getPageSize());
 
 		Sort sort = new Sort(new SortField("id", SortField.Type.LONG, true));
